+++ app/Http/Controllers/Auth/DashBoardController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\VideoResource;
+use App\Models\Folder;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
@@ -13,11 +14,19 @@ class DashBoardController extends Controller
     {
         /** @var \App\Models\User $user */
         $user = auth()->user();
+
+        /** @var ?Folder $folder */
+        $folder = $request->has('folder')
+            ? Folder::findBySlug($request->get('folder'))
+            : null;
+
+        if ($folder) $folder->load('parent');
         
         $assetPath = asset(config('thumbnail.asset'));
-        $totalVideos = $user->totalVideos();
-        $usedVideos = $user->loadCount('videos')->videos_count;
-        $videos = $user->videos()->with('thumbnail')
+        $totalVideos = fn() => $user->totalVideos();
+        $usedVideos = fn() => $user->loadCount('videos')->videos_count;
+        $videos = fn() => $user->videos()->with('thumbnail')
+            ->where('folder_id', $folder?->id)
             ->limit(10)
             ->get()
             ->map(function ($v) use($assetPath) {
@@ -26,7 +35,11 @@ class DashBoardController extends Controller
             })
             ->toArray();
 
-        return Inertia::render('Dashboard', compact('usedVideos', 'totalVideos', 'videos'));
+        $folders = fn() => $user->folders()
+            ->where('parent_id', $folder?->id)->get()
+            ->append('date');
+
+        return Inertia::render('Dashboard', compact('usedVideos', 'totalVideos', 'videos', 'folder', 'folders'));
     }
 
 }