+++ resources/js/helpers/index.js
@@ -1,0 +1,11 @@
+
+export function slugify(string, separator="-") {
+    return string
+        .toString()
+        .normalize('NFD')
+        .replace(/[\u0300-\u036f]/g, '')
+        .toLowerCase()
+        .trim()
+        .replace(/[^a-z0-9 ]/g, '')
+        .replace(/\s+/g, separator);
+}