+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## releases
 
 ### Features
+ - Add Folder System [#132](https://github.com/AmolKumarGupta/Websnapper/issues/132)
  - Uploading Video in chunks [#122](https://github.com/AmolKumarGupta/Websnapper/issues/122)
  - Delete Video permanently [#108](https://github.com/AmolKumarGupta/Websnapper/issues/108)
  - Add thumbnail for videos [#40](https://github.com/AmolKumarGupta/Websnapper/issues/40)